feat: add negative sampler and  MultiTaskDatasetManager
